requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.0
websockets>=10.0
python-multipart
jinja2
passlib[bcrypt]
python-jose[cryptography]